// server/src/domain/enums/ExamTypeEnum.ts
export enum ExamTypeEnum {
  HEMATOLOGIA = "HEMATOLOGIA",
  BIOQUIMICA_CLINICA = "BIOQUIMICA_CLINICA",
  IMUNOLOGIA = "IMUNOLOGIA",
  URINA = "URINA",
  FEZES = "FEZES",
  MICROBIOLOGIA = "MICROBIOLOGIA",
  HORMONIOS = "HORMONIOS",
  PATOLOGIA = "PATOLOGIA",
  BIOLOGIA_MOLECULAR = "BIOLOGIA_MOLECULAR",
  CARDIOLOGIA = "CARDIOLOGIA",
  ULTRASSONOGRAFIA = "ULTRASSONOGRAFIA",
  DERMATOLOGIA = "DERMATOLOGIA",
}